#10 Refactored packages according to module group inside  launchpad-s emu-web module
