+++ launchpad-s/emu-web/src/main/kotlin/io/lp4k/launchpad/s/emulator/web/input/EventBusMessageDebugger.kt
@@ -14,7 +14,7 @@
  *    limitations under the License.
  *
  */
-package io.lp4k.emulator.input
+package io.lp4k.launchpad.s.emulator.web.input
 
 import io.vertx.core.Handler
 import io.vertx.ext.bridge.BridgeEventType
