+++ launchpad-s/emu-web/src/test/kotlin/io/lp4k/launchpad/s/emulator/web/input/EventBusMessageDebuggerTest.kt
@@ -14,7 +14,7 @@
  *    limitations under the License.
  *
  */
-package io.lp4k.emulator.input
+package io.lp4k.launchpad.s.emulator.web.input
 
 import io.vertx.ext.bridge.BridgeEventType
 import org.junit.jupiter.api.BeforeEach
